Move ZoomToDropdown logic to its own factory
