+++ src/app/factories/index.js
@@ -1,0 +1,10 @@
+
+import angular from 'angular';
+
+import {ZoomToDropdownFactory} from './zoomtodropdown.factory';
+
+export const adbFactoriesModule = 'adb.factories';
+
+angular
+  .module(adbFactoriesModule, [])
+  .factory('ZoomToDropdown', ZoomToDropdownFactory);
